[Paymium] - Add getFunding method
